Update build and project configuration
