+++ MANIFEST.in
@@ -13,12 +13,8 @@ global-exclude __pycache__/**
 
 # Top-level Config
 include CHANGELOG
-include classifiers.txt
 include LICENSE.txt
 include MANIFEST.in
-include setup.cfg
-include versioneer.py
-include _setup_support.py
 
 # Package files and data
 include bokeh/LICENSE.txt
