+++ ci/install_bokeh_package.sh
@@ -2,7 +2,7 @@
 
 set -x #echo on
 
-conda install --yes --quiet --offline /tmp/conda-bld/noarch/bokeh-*.bz2
+$CONDA/bin/conda install --yes --quiet --offline /tmp/conda-bld/noarch/bokeh-*.bz2
 
 # TODO (bev) remove this when move from bokeh -> src dir struture
-python setup.py --install-js
+rm -rf bokeh
