--- classifiers.txt
@@ -1,27 +1,0 @@
-Development Status :: 5 - Production/Stable
-Environment :: Console
-Environment :: Web Environment
-Intended Audience :: Developers
-Intended Audience :: Education
-Intended Audience :: End Users/Desktop
-Intended Audience :: Financial and Insurance Industry
-Intended Audience :: Healthcare Industry
-Intended Audience :: Information Technology
-Intended Audience :: Legal Industry
-Intended Audience :: Other Audience
-Intended Audience :: Science/Research
-License :: OSI Approved :: BSD License
-Programming Language :: Python
-Programming Language :: Python :: 3
-Programming Language :: Python :: 3 :: Only
-Programming Language :: Python :: 3.8
-Programming Language :: Python :: 3.9
-Programming Language :: Python :: 3.10
-Programming Language :: JavaScript
-Topic :: Office/Business
-Topic :: Office/Business :: Financial
-Topic :: Scientific/Engineering
-Topic :: Scientific/Engineering :: Visualization
-Topic :: Scientific/Engineering :: Mathematics
-Topic :: Scientific/Engineering :: Information Analysis
-Topic :: Utilities
