--- conda.recipe/bld.bat
@@ -1,2 +1,0 @@
-python setup.py --quiet install --build-js --single-version-externally-managed --record=record.txt
-if errorlevel 1 exit 1
