+++ conda.recipe/meta.yaml
@@ -1,14 +1,18 @@
-{% set data = load_setup_py_data() %}
+{% set data = load_file_data('../pyproject.toml', from_recipe_dir=True) %}
 
 package:
   name: bokeh
-  version: {{ data['version'] }}
+  version: {{ VERSION }}
 
 source:
-  path: ..
+  url: ../dist/bokeh-{{ VERSION }}-py3-none-any.whl
 
 build:
   noarch: python
+  script:
+     - {{ PYTHON }} -m pip install -vv --no-deps bokeh-{{ VERSION }}-py3-none-any.whl
+  script_env:
+    - VERSION
   entry_points:
     - bokeh = bokeh.__main__:main
 
@@ -19,17 +23,13 @@ extra:
 
 requirements:
   build:
-    - contourpy >=1
     - jinja2
-    - nodejs 16.*
-    - numpy
     - packaging
     - python
     - python-dateutil
     - pyyaml
     - ripgrep
-    - setuptools
-    - tornado >=5
+    - setuptools >=62.3.3
     - yaml
 
   host:
@@ -38,19 +38,14 @@ requirements:
   run:
     - python >=3.8
 
-    {% for dep in data['install_requires'] %}
+    {% for dep in data['project']['dependencies'] %}
     - {{ dep.lower() }}
     {% endfor %}
 
 about:
-  home: https://www.bokeh.org
-  license: {{ data['license'] }}
-  summary: {{ data['description'] }}
-  description: |
-    Bokeh is a Python library for creating interactive visualizations
-    for modern web browsers. It helps you build beautiful graphics,
-    ranging from simple plots to complex dashboards with streaming
-    datasets. With Bokeh, you can create JavaScript-powered visualizations
-    without writing any JavaScript yourself.
-  doc_url: https://docs.bokeh.org
-  dev_url: {{ data['url'] }}
+  license: BSD-3-Clause
+  summary: {{ data['project']['description'] }}
+  description: {{ data['project']['readme'] }}
+  home: {{ data['project']['urls']['homepage'] }}
+  doc_url: {{ data['project']['urls']['documentation'] }}
+  dev_url: {{ data['project']['urls']['repository'] }}
