--- tests/.coveragerc
@@ -1,2 +1,0 @@
-[run]
-omit = *_testing/*, *sphinxext/*, */_version.py
